updating readme and moving mock to helpers folder again
