+++ tests/helpers/google-place-autocomplete-response-mock.js
